[3.3.5a][ICC/Professor Putricide] Volatile Ooze Eruption fix
by @ManuFe and @Rushor; closes #12336

In the existing source, the rooting aura is removed after the explosion,
therefore the knockback effect does not work because the player is still rooted.
By swapping these 2 lines, the aura is removed, then the explosion happens.

Result: rooting is removed before the knockback effect is cast on the player.