Core/authserver.conf.dist: missing description line (#17038)
This line is missing from authserver.conf.dist :
`#                     MySQL server and their own thread on the MySQL server.`

I have copied this description text line from the matching description in
`worldserver.conf.dist` (line 108 in the 3.3.5 worldserver source file).

This line is also missing in bnetserver.conf.dist , but the files are different
in line number, filename and folder, so bnetserver needs a separate commit.